# Conflicts:
#	app/Http/Controllers/AuthenticationController.php
#	app/Models/User.php
#	vendor/composer/autoload_classmap.php

+++ app/Http/Controllers/Events.php
@@ -9,10 +9,8 @@ use App\Http\Controllers\Controller;
 
 class Events extends Controller
 {
-    //
     /**
-     * Display a listing of the resource.
-     *
+     * Display a listing of all events in the database.
      * @return Response
      */
     public function index($id = null) {
@@ -24,63 +22,75 @@ class Events extends Controller
     }
 
     /**
+     * Display a listing of events given the conferenceID.
+     * @return Response
+     */
+    public function getEventByConferenceID($conferenceID) {
+    $event = Event::where('conferenceID',$conferenceID)->get();
+      if (count($event) == 0) {return response("No events for conferenceID {$conferenceID}.", 404);}
+      return $event;
+    }
+
+    /**
      * Store a newly created resource in storage.
-     *
      * @param  Request  $request
      * @return Response
-     *('id', 'eventName', 'date','location','time','seatsCount','conferenceID');
      */
     public function store(Request $request) {
         $event = new Event;
         $event->eventName = $request->input('eventName');
         $event->date = $request->input('date');
         $event->location = $request->input('location');
-        $event->time = $request->input('time');
-        $event->seatsCount = $request->input('seatsCount');
+        $event->startTime = $request->input('startTime');
+        $event->endTime = $request->input('endTime');
+        $event->capacity = $request->input('capacity');
+        $event->description = $request->input('description');
         $event->conferenceID = $request->input('conferenceID');
         $event->save();
-        return response()->json(['Employee ID' => $event->id]);
+        return response()->json(['id' => $event->id]);
     }
 
     /**
-     * Display the specified resource.
-     *
+     * Display the event given the eventID.
      * @param  int  $id
      * @return Response
      */
     public function show($id) {
-        return Event::find($id);
+      $event = Event::find($id);
+        if (count($event) == 0) {return response("No event for id {$id}.", 404);}
+        return $event;
     }
 
     /**
-     * Update the specified resource in storage.
-     *
+     * Update the specified event give the eventID.
      * @param  Request  $request
      * @param  int  $id
      * @return Response
      */
     public function update(Request $request, $id) {
         $event = Event::find($id);
-
+        if (is_null($event)) {return response("No event for id {$id}.", 404);}
         $event->eventName = $request->input('eventName');
         $event->date = $request->input('date');
         $event->location = $request->input('location');
-        $event->time = $request->input('time');
-        $event->seatsCount = $request->input('seatsCount');
+        $event->startTime = $request->input('startTime');
+        $event->endTime = $request->input('endTime');
+        $event->capacity = $request->input('capacity');
+        $event->description = $request->input('description');
         $event->conferenceID = $request->input('conferenceID');
         $event->save();
-        return response()->json(['Employee ID' => $event->id]);
+        return response()->json(['id' => $event->id]);
     }
 
     /**
-     * Remove the specified resource from storage.
-     *
+     * Remove the event given the eventID.
      * @param  int  $id
      * @return Response
      */
     public function destroy($id) {
         $event = Event::find($id);
+        if (is_null($event)) {return response("No event for id {$id}.", 404);}
         $event->delete();
-        return response()->json(['Employee ID' => $event->id]);
+        return response()->json(['id' => $event->id]);
     }
 }
